Updated services and integrated UI

+++ src/main/java/com/portfolio/management/app/repository/LotRepository.java
@@ -7,7 +7,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import com.portfolio.management.app.entity.Lot;
 import com.portfolio.management.app.entity.OwnedStock;
 
-public interface LotRepository extends JpaRepository<Lot, Integer> {
+public interface LotRepository extends JpaRepository<Lot, Long> {
 
 	List<Lot> findLotByOwnedStock(OwnedStock ownedStock);
 }
